[chore] #48 패키지 변경에 따른 변경 사항 적용
