+++ src/main/java/com/ggang/be/api/gongbaekTimeSlot/service/GongbaekTimeSlotService.java
@@ -1,7 +1,7 @@
 package com.ggang.be.api.gongbaekTimeSlot.service;
 
-import com.ggang.be.domain.gongbaekTimeSlot.GongbaekTimeSlotEntity;
-import com.ggang.be.domain.gongbaekTimeSlot.dto.GongbaekTimeSlotRequest;
+import com.ggang.be.domain.timslot.gongbaekTimeSlot.GongbaekTimeSlotEntity;
+import com.ggang.be.domain.timslot.gongbaekTimeSlot.dto.GongbaekTimeSlotRequest;
 import com.ggang.be.domain.user.UserEntity;
 
 public interface GongbaekTimeSlotService {
