+++ src/main/java/com/ggang/be/api/group/dto/RegisterGongbaekRequest.java
@@ -3,9 +3,9 @@ package com.ggang.be.api.group.dto;
 import com.ggang.be.api.facade.GroupType;
 import com.ggang.be.domain.constant.Category;
 import com.ggang.be.domain.constant.WeekDate;
-import com.ggang.be.domain.gongbaekTimeSlot.dto.GongbaekTimeSlotRequest;
+import com.ggang.be.domain.timslot.gongbaekTimeSlot.dto.GongbaekTimeSlotRequest;
 import com.ggang.be.domain.group.dto.RegisterGroupServiceRequest;
-import com.ggang.be.domain.lectureTimeSlot.dto.LectureTimeSlotRequest;
+import com.ggang.be.domain.timslot.lectureTimeSlot.dto.LectureTimeSlotRequest;
 import com.ggang.be.domain.user.UserEntity;
 import java.time.LocalDate;
 import java.util.Objects;
