+++ src/main/java/com/ggang/be/api/user/vo/TimeTableVo.java
@@ -1,7 +1,7 @@
 package com.ggang.be.api.user.vo;
 
 import com.ggang.be.domain.constant.WeekDate;
-import com.ggang.be.domain.lectureTimeSlot.vo.LectureTimeSlotVo;
+import com.ggang.be.domain.timslot.lectureTimeSlot.vo.LectureTimeSlotVo;
 
 public record TimeTableVo(WeekDate weekDay,
                           Double startTime,
