+++ src/main/java/com/ggang/be/domain/group/dto/RegisterGroupServiceRequest.java
@@ -3,7 +3,6 @@ package com.ggang.be.domain.group.dto;
 import com.ggang.be.api.facade.GroupType;
 import com.ggang.be.domain.constant.Category;
 import com.ggang.be.domain.constant.WeekDate;
-import com.ggang.be.domain.gongbaekTimeSlot.GongbaekTimeSlotEntity;
 import com.ggang.be.domain.user.UserEntity;
 import java.time.LocalDate;
