+++ src/main/java/com/ggang/be/domain/group/everyGroup/dto/EveryGroupDto.java
@@ -1,6 +1,6 @@
 package com.ggang.be.domain.group.everyGroup.dto;
 
-import com.ggang.be.domain.gongbaekTimeSlot.GongbaekTimeSlotEntity;
+import com.ggang.be.domain.timslot.gongbaekTimeSlot.GongbaekTimeSlotEntity;
 import com.ggang.be.domain.group.everyGroup.EveryGroupEntity;
 import com.ggang.be.domain.user.UserEntity;
