+++ src/main/java/com/ggang/be/domain/timslot/gongbaekTimeSlot/GongbaekTimeSlotEntity.java
@@ -1,5 +1,4 @@
-package com.ggang.be.domain.lectureTimeSlot;
-
+package com.ggang.be.domain.timslot.gongbaekTimeSlot;
 
 import com.ggang.be.domain.BaseTimeEntity;
 import com.ggang.be.domain.constant.WeekDate;
@@ -15,16 +14,16 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import lombok.*;
 
-@Entity(name = "lecture_time_slot")
+@Entity(name = "gongbaek_time_slot")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LectureTimeSlotEntity extends BaseTimeEntity {
+public class GongbaekTimeSlotEntity extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false)
-    @Enumerated(value = EnumType.STRING)
+    @Enumerated(EnumType.STRING)
     private WeekDate weekDate;
 
     private double startTime;
@@ -36,7 +35,7 @@ public class LectureTimeSlotEntity extends BaseTimeEntity {
 
 
     @Builder
-    private LectureTimeSlotEntity(WeekDate weekDate, double startTime, double endTime,
+    private GongbaekTimeSlotEntity(WeekDate weekDate, double startTime, double endTime,
         UserEntity userEntity) {
         this.weekDate = weekDate;
         this.startTime = startTime;
