+++ src/main/java/com/ggang/be/domain/timslot/gongbaekTimeSlot/dto/GongbaekTimeSlotRequest.java
@@ -1,4 +1,4 @@
-package com.ggang.be.domain.gongbaekTimeSlot.dto;
+package com.ggang.be.domain.timslot.gongbaekTimeSlot.dto;
 
 import com.ggang.be.domain.constant.WeekDate;
 import com.ggang.be.domain.user.UserEntity;
