+++ src/main/java/com/ggang/be/domain/timslot/gongbaekTimeSlot/infra/GongbaekTimeSlotRepository.java
@@ -1,7 +1,7 @@
-package com.ggang.be.domain.gongbaekTimeSlot.infra;
+package com.ggang.be.domain.timslot.gongbaekTimeSlot.infra;
 
 import com.ggang.be.domain.constant.WeekDate;
-import com.ggang.be.domain.gongbaekTimeSlot.GongbaekTimeSlotEntity;
+import com.ggang.be.domain.timslot.gongbaekTimeSlot.GongbaekTimeSlotEntity;
 import com.ggang.be.domain.user.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
