+++ src/main/java/com/ggang/be/domain/user/UserEntity.java
@@ -3,8 +3,8 @@ package com.ggang.be.domain.user;
 import com.ggang.be.domain.BaseTimeEntity;
 import com.ggang.be.domain.constant.Gender;
 import com.ggang.be.domain.constant.Mbti;
-import com.ggang.be.domain.gongbaekTimeSlot.GongbaekTimeSlotEntity;
-import com.ggang.be.domain.lectureTimeSlot.LectureTimeSlotEntity;
+import com.ggang.be.domain.timslot.gongbaekTimeSlot.GongbaekTimeSlotEntity;
+import com.ggang.be.domain.timslot.lectureTimeSlot.LectureTimeSlotEntity;
 import com.ggang.be.domain.school.SchoolEntity;
 import com.ggang.be.domain.userEveryGroup.UserEveryGroupEntity;
 import com.ggang.be.domain.userOnceGroup.UserOnceGroupEntity;
