+++ src/test/java/com/ggang/be/domain/lectureTimeSlot/application/LectureTimeSlotServiceImplTest.java
@@ -6,9 +6,10 @@ import static org.mockito.Mockito.times;
 import com.ggang.be.domain.constant.Gender;
 import com.ggang.be.domain.constant.Mbti;
 import com.ggang.be.domain.constant.WeekDate;
-import com.ggang.be.domain.lectureTimeSlot.LectureTimeSlotEntity;
-import com.ggang.be.domain.lectureTimeSlot.vo.LectureTimeSlotVo;
-import com.ggang.be.domain.lectureTimeSlot.infra.LectureTimeSlotRepository;
+import com.ggang.be.domain.timslot.lectureTimeSlot.LectureTimeSlotEntity;
+import com.ggang.be.domain.timslot.lectureTimeSlot.application.LectureTimeSlotServiceImpl;
+import com.ggang.be.domain.timslot.lectureTimeSlot.vo.LectureTimeSlotVo;
+import com.ggang.be.domain.timslot.lectureTimeSlot.infra.LectureTimeSlotRepository;
 import com.ggang.be.domain.school.SchoolEntity;
 import com.ggang.be.domain.user.UserEntity;
 import java.util.List;
